+++ CHANGELOG.md
@@ -4,14 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.4.0] - 2025-11-XX
+## [1.4.0] - 2025-11-13
 
 ### Added
-- Added the capability for inline comments of rule violations across a PR, along with a more detailed summary comment for new issues vs technical debt
+- Added the capability for inline comments of rule violations across a PR, with details from the json output
 - Allowed multiple engines/tags to be passed in to the scanner (as per SFCA v5.6.1 release capabilities, instead of custom logic)
 - Allowed multiple output file types to be passed in to the scanner
 ### Changed
-- Updated the existing summary comment to use markdown instead
+- Updated the existing summary comment to use markdown instead and provide extra tech debt vs new issue detail along with severity information
 ### Fixed 
 -